BitmapSurface: rework selection
The last implementation didn't work properly as I forgot to take into account that we need to know the new tile positions before we can draw on them, that's why the selected content would (when re-drawn...) offset to match topLeft of the tile pos. This has now been rewritten so it works properly.