Back to floodfill for bitmapimage
This will be one hella of a ride... I basically came to the conclusion that most of the work i've made over the past month has been for nothing... I'm working on reversing all my bitmapsurface implementation and going back to bitmapimage.

On the positive side:
This should make the implementation of mypaint much faster, as I don't have to rewrite anything anymore...
On the negative side:
All that time wasted... but gotta kill that baby.