Improve polyline performance.
Since the polyline is a floating stroke and therefore has to clear the surface each time it is drawn, the performance ends up being pretty terrible. To minimize this problem, we'll only keep polyline in the mypaint buffer and draw the rest of the image below as a cache from the layer image, then only apply the finished polyline to the image again. This ensures that the stroke stays fast, no matter the size of the image.