Fix canvas not being updated properly when starting new stroke...
The solution here was to always paint the buffered image underneath the tiles, making sure that when a paintEvent is called, the screen should update with the content of the persisted bitmapimage. To avoid occluding the frame with the previous opaque image, we simply get the coordinate of the tile and clear the painter before drawing the tile. The only exception here is polyline, which required special handling to keep it fast.

+ Fixing white refresh artifacts
+ Fixed incorrect canvas cursor size.

Known issue:
We rely on mypaint state now to get the actual brush size.. this might be a problem because the state is only updated when drawing, so whenever you change the brush size, the old cursor will be shown until you've drawn.