Introduce SAVESTATE_ID to deal with potential corrupt save state
The mCurrentState was flawed in the way that it assumed you would always have a createState followed by record(...) which is not always the case...

the scenario i ran into was:
creating a selection and then moving some keyframes, leaving the scrubbed position with no keyframe, this would then call createState, deleting the current mCurrentState in the manager (and leaving the one stored in MoveTool dangling) which when you then tried to apply the selection, it would crash because the UndoSaveState had been deleted memory wise but the ptr wouldn't know that.

The solution i came up with to fix this is that we instead of getting access to the raw ptr, we get a handle now called SAVESTATE_ID. The id is used to fetch a given saveState stored temporarily in undoRedoManager. This now also means that we can create multiple save states and apply all of them without having to think about dealing with their lifetime.

In order to still allow the caller to add additional state to the record, I have also introduced a new struct called UserSaveState which is meant to work as a mega struct in the way that it can store all kinds of data but you should only add the data you actually need.